+++ R/simulate.r
@@ -10,8 +10,19 @@ sql_escape_ident.DBITestConnection <- function(con, x) {
   sql_quote(x, "`")
 }
 
+#' @export
+sql_subquery.DBITestConnection <- function(con, from, name = unique_name(), ...) {
+  if (is.ident(from)) {
+    setNames(from, name)
+  } else {
+    build_sql("(", from, ") ", ident(name %||% random_table_name()), con = con)
+  }
+}
+
 # DBI connections --------------------------------------------------------------
 
+#' @export
+#' @rdname tbl_lazy
 simulate_dbi <- function() {
   structure(
     list(),
@@ -19,6 +30,8 @@ simulate_dbi <- function() {
   )
 }
 
+#' @export
+#' @rdname tbl_lazy
 simulate_sqlite <- function() {
   structure(
     list(),
@@ -26,9 +39,24 @@ simulate_sqlite <- function() {
   )
 }
 
+#' @export
+#' @rdname tbl_lazy
 simulate_postgres <- function() {
   structure(
     list(),
     class = c("PostgreSQLConnection", "DBIConnection")
   )
 }
+
+#' @export
+#' @rdname tbl_lazy
+simulate_mssql <- function() {
+  structure(
+    list(),
+    class = c("Microsoft SQL Server", "DBITestConnection")
+  )
+
+}
+
+
+