+++ R/sql-generic.R
@@ -1,3 +1,17 @@
+#' More SQL generics
+#'
+#' These are new, so not included in dplyr for backward compatibility
+#' purposes.
+#'
+#' @keywords internal
+#' @export
+sql_escape_logical <- function(con, x) {
+  UseMethod("sql_escape_logical")
+}
+
+
+# DBIConnection methods -----------------------------------------------------------------
+
 #' @export
 sql_select.DBIConnection <- function(con, select, from, where = NULL,
                                group_by = NULL, having = NULL,
@@ -255,4 +269,3 @@ sql_translate_env.DBIConnection <- function(con) {
     base_win
   )
 }
-