+++ R/tbl-lazy.R
@@ -1,9 +1,24 @@
-tbl_lazy <- function(df) {
-  make_tbl("lazy", ops = op_base_local(df))
+#' Create a local lazy tibble
+#'
+#' These functions are useful for testing SQL generation without having to
+#' have an active database connection.
+#'
+#' @keywords internal
+#' @export
+#' @examples
+#' library(dplyr)
+#' df <- data.frame(x = 1, y = 2)
+#'
+#' df_sqlite <- tbl_lazy(df, src = simulate_sqlite())
+#' df_sqlite %>% summarise(x = sd(x)) %>% show_query()
+tbl_lazy <- function(df, src = NULL) {
+  make_tbl("lazy", ops = op_base_local(df), src = src)
 }
 
-lazy_frame <- function(...) {
-  tbl_lazy(data_frame(...))
+#' @export
+#' @rdname tbl_lazy
+lazy_frame <- function(..., src = NULL) {
+  tbl_lazy(tibble(...), src = src)
 }
 
 #' @export
@@ -38,7 +53,7 @@ print.tbl_lazy <- function(x, ...) {
 
 # Single table methods ----------------------------------------------------
 
-#' @export
+# registered onLoad
 filter.tbl_lazy <- function(.data, ...) {
   dots <- quos(...)
   dots <- partial_eval(dots, vars = op_vars(.data))
@@ -61,7 +76,7 @@ arrange.tbl_lazy <- function(.data, ...) {
 }
 #' @export
 arrange_.tbl_lazy <- function(.data, ..., .dots = list()) {
-  dots <- dplyr:::compat_lazy_dots(dots, caller_env(), ...)
+  dots <- dplyr:::compat_lazy_dots(.dots, caller_env(), ...)
   arrange(.data, !!! dots)
 }
 
@@ -91,7 +106,7 @@ rename_.tbl_lazy <- function(.data, ..., .dots = list()) {
 
 #' @export
 summarise.tbl_lazy <- function(.data, ...) {
-  dots <- quos(...)
+  dots <- quos(..., .named = TRUE)
   add_op_single("summarise", .data, dots = dots)
 }
 #' @export