+++ R/tbl-sql.r
@@ -12,10 +12,8 @@
 #'   dplyr. However, you should usually be able to leave this blank and it
 #'   will be determined from the context.
 tbl_sql <- function(subclass, src, from, ..., vars = NULL) {
-  # If not literal ql, must be a table identifier
-  if (!is.sql(from)) {
-    from <- ident(from)
-  }
+  # If not literal sql, must be a table identifier
+  from <- as.sql(from)
 
   vars <- db_query_fields(src$con, from)
   ops <- op_base_remote(from, vars)
@@ -80,8 +78,15 @@ print.tbl_sql <- function(x, ..., n = NULL, width = NULL) {
 }
 
 #' @export
+glimpse.tbl_sql <- function(.data, width = NULL, n = 25, ...) {
+  .data <- collect(head(.data, n = n))
+  glimpse(.data, width = width, ...)
+}
+
+#' @export
 pull.tbl_sql <- function(.data, var = -1) {
-  var <- dplyr:::find_var(var, tbl_vars(.data))
+  expr <- enquo(var)
+  var <- dplyr:::find_var(expr, tbl_vars(.data))
 
   .data <- select(.data, !! sym(var))
   .data <- collect(.data)
@@ -281,11 +286,11 @@ anti_join.tbl_lazy <- function(x, y, by = NULL, copy = FALSE,
 
 # Set operations ---------------------------------------------------------------
 
-#' @export
+# registered onLoad
 intersect.tbl_lazy <- function(x, y, copy = FALSE, ...) {
   add_op_set_op(x, y, "INTERSECT", copy = copy, ...)
 }
-#' @export
+# registered onLoad
 union.tbl_lazy <- function(x, y, copy = FALSE, ...) {
   add_op_set_op(x, y, "UNION", copy = copy, ...)
 }
@@ -293,7 +298,7 @@ union.tbl_lazy <- function(x, y, copy = FALSE, ...) {
 union_all.tbl_lazy <- function(x, y, copy = FALSE, ...) {
   add_op_set_op(x, y, "UNION ALL", copy = copy, ...)
 }
-#' @export
+# registered onLoad
 setdiff.tbl_lazy <- function(x, y, copy = FALSE, ...) {
   add_op_set_op(x, y, "EXCEPT", copy = copy, ...)
 }